+++ src/main/webapp/mypage/noticeDetail.jsp
@@ -41,9 +41,9 @@
             border: none;
         }
 
-        div {
-            border: 1px solid #fff;
-        }
+/*         div { */
+/*             border: 1px solid #fff; */
+/*         } */
 
         body {
             width: 100vw;
@@ -55,8 +55,7 @@
         .container {
             padding: 0 75px;
             height: 100%;
-            background-color: bisque;
-            overflow-y: scroll;
+/*             background-color: bisque; */
             width: 100vw;
     		margin-bottom:200px;
         }
@@ -66,45 +65,45 @@
             font-weight: var(--font-weight);
             line-height: var(--line-height);
             margin-bottom: 2vw;
-            background-color: #00000d50;
+/*             background-color: #00000d50; */
         }
 
         /*qna css */
         .qna {
             height: 100%;
             margin-top: 8vh;
-            background-color: lightgray;
+/*             background-color: lightgray; */
         }
 
         .qnalist {
             width: 100%;
-            height: 65vh;
+            height: 48vh;
             margin-top: 5vh;
-            background-color: #00000d80;
+/*             background-color: #00000d80; */
         }
 
         #category {
             width: 100%;
-            font-size: 20px;
+            font-size: 18px;
             font-weight: var(--font-weight);
             line-height: var(--line-height);
         }
 
         #title {
             width: 100%;
-            font-size: 20px;
+            font-size: 18px;
             font-weight: var(--font-weight);
             line-height: var(--line-height);
         }
         #writer{
         	width: 100%;
-            font-size: 20px;
+            font-size: 18px;
             font-weight: var(--font-weight);
             line-height: var(--line-height);
         }
         #contents{
             width: 100%;
-            font-size: 20px;
+            font-size: 18px;
             font-weight: var(--font-weight);
             line-height: var(--line-height);
         }
@@ -113,12 +112,11 @@
         #listBtn {
             margin-top: 2vh;
             width: 6vw;
-            height: 40px;
+            height: 35px;
             text-align: center;
             border-radius: 5px;
-            font-size: 18px;
+            font-size: 15px;
             font-weight: var(--font-weight);
-            line-height: var(--line-height);
             border: 1px solid gray;
             cursor: pointer;
             background: var(--background-color);
